move check binary/text to utils function
